Move visit_dot back to generator.cc
