+++ lib/codegen/analysis/swizzle.cc
@@ -40,31 +40,10 @@ void swizzle::run(ir::module &) {
         else
           vec_[layout] = 2*layouts_->get(mma_dot_b)->to_mma()->rep(1);
       }
-      else{
-        // if (mma_dot_a) {
-        //   per_phase_[layout] = std::max<int>(128 / (in_layout->mts(ord[0])*in_layout->nts(ord[0])*dtsize), 1);
-        //   max_phase_[layout] = 4 / per_phase_[layout];
-        //   vec_[layout]       = 8; // FIXME!!!!!!!!!!!!!!!!!
-        // }
-        // if (mma_dot_b) {
-        //   per_phase_[layout] = std::max<int>(128 / (in_layout->mts(ord[0])*in_layout->nts(ord[0])*dtsize), 1);
-        //   max_phase_[layout] = 4 / per_phase_[layout];
-        //   vec_[layout]       = 8; // FIXME!!!!!!!!!!!!!!!!!
-        // } 
-        // else 
-        {
-        
+      else {
           per_phase_[layout] = std::max<int>(128 / (in_layout->mts(ord[0])*in_layout->nts(ord[0])*dtsize), 1);
-          // max_phase_[layout] = 8 / per_phase_[layout];
           max_phase_[layout] = layout->get_mma_strided() / per_phase_[layout];
-          vec_[layout]       = layout->get_mma_vec(); // FIXME!!!!!!!!!!!!!!!!!
-          // std::cout << "mma swizzle\n"
-          //           << "mma_dot_a: " << mma_dot_a << "\n"
-          //           << "mts[0]: " << in_layout->mts(ord[0]) << "\tnts[0]: " << in_layout->nts(ord[0]) << "\tdtsize: " << dtsize << "\n"
-          //           << "per_phase_: " << per_phase_[layout] << "\n"
-          //           << "max_phase_: " << max_phase_[layout] << "\n"
-          //           << "vec_: " << vec_[layout] << "\n\n";
-        }
+          vec_[layout]       = layout->get_mma_vec();
       }
     }
 }
