+++ lib/codegen/transform/prefetch.cc
@@ -29,11 +29,11 @@ void prefetch::run(ir::module &mod) {
   std::vector<ir::dot_inst*> to_prefetch;
   ir::for_each_instruction(mod, [&](ir::instruction *i) {
     if (auto *dot = dynamic_cast<ir::dot_inst*>(i)) {
-      // Now only do prefetching when dot is fp16
-      // TODO: should check dot layout here.
+      // Now only do prefetching when dot is using tensor cores
       if (!(dot->get_operand(0)->get_type()->get_scalar_ty()->is_fp16_ty() ||
             dot->get_operand(0)->get_type()->get_scalar_ty()->is_bf16_ty() ||
-            (dot->get_operand(0)->get_type()->get_scalar_ty()->is_fp32_ty() && dot->allow_tf32())
+            (dot->get_operand(0)->get_type()->get_scalar_ty()->is_fp32_ty() && dot->allow_tf32()
+             && tgt_->as_nvidia() && tgt_->as_nvidia()->sm() >= 80)
            )
          )
         return;
